services/ai-agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
cassandra-driver==3.28.0
httpx==0.26.0
langgraph==0.0.20
openai==1.12.0